requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
tensorflow-cpu==2.13.0
numpy==1.24.3
pillow==10.2.0
python-multipart==0.0.9





# fastapi==0.110.0
# uvicorn==0.27.1
# tensorflow==2.15.0
# numpy = 1.26.4 
# pillow==10.2.0
# scipy==1.10.1

# scikit-learn==1.3.2
# matplotlib==3.7.3
# seaborn==0.12.2
#  fastapi==0.110.0 
#  uvicorn==0.27.1
